+++ internal/constants/constants.go
@@ -2,4 +2,6 @@ package constants
 
 const (
 	ISO8601DateFormat = "2006-01-02T15:04:05-0700"
+	MaxPaymentAddresses = 1
+	MaxStakeAddresses = 1
 )